requirements.txt
flask
tensorflow-cpu
numpy
google-generativeai
gunicorn
google-cloud-texttospeech
